% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arpchsurv.R
\name{arpchsurv}
\alias{arpchsurv}
\title{The adaptive ridge and ridge estimates from the piecewise constant hazard model}
\usage{
arpchsurv(time, status, cuts = NULL, weights = rep(1, length(time)),
  pen = exp(seq(log(0.1), log(1000), length = 100)), a = rep(0,
  length(cuts) + 1), AR = TRUE, ...)
}
\arguments{
\item{time}{the observed time data.}

\item{status}{status of the time data. TRUE for true time and FALSE for censored time.}

\item{cuts}{a sequence of cuts. Default to NULL which corresponds to the exponential model.}

\item{weights}{an optional weight sequence. Default to 1. Should be of the size of time and status.}

\item{pen}{a sequence of penalty values.}

\item{a}{the log-hazard value at the initialization step. Default to the unpenalized log-hazard estimator.}

\item{AR}{do you want to compute the adaptive-ridge? If FALSE compute the ridge estimator. Default to TRUE.}

\item{w}{the w sequence values for the adaptive ridge algorithm at the initialization step. Default to 1. Should be of the size of the cuts.}

\item{tol}{the tolerance parameter for convergence of the algorithm. Default to 1e-7.}

\item{itermax}{the maximum number of iterations. If the algorithm has not converged before itermax iterations then the algorithm exits the program. Default to 1e+5.}

\item{CI}{should the confidence intervals be computed? Default to TRUE.}

\item{alphaCI}{the value of alpha for 1-alpha confidence intervals. Default to 0.05.}

\item{logtransf}{should the confidence intervals be computed using the log-transformtation? Default to TRUE.}
}
\value{
\tabular{lll}{
\code{res} \tab  \code{ } \tab the list of adaptive-ridge estimators for each penalty value\cr
\tab \tab \cr
\code{final.a} \tab \code{ }  \tab the final log-hazard that minimizes the BIC\cr
\tab \tab \cr
\code{final.cuts} \tab  \code{ } \tab the final cut vector found from the BIC\cr
\tab \tab \cr
\code{bic} \tab \code{ }  \tab the minimal value of the BIC\cr
\tab \tab \cr
\code{pos} \tab  \code{ } \tab the position of the penalty that minimizes the BIC\cr
\tab \tab \cr
\code{CIleft} \tab  \code{ } \tab the left confidence intervals for the final hazard\cr
\tab \tab \cr
\code{CIright} \tab \code{ }  \tab the right confidence intervals for the final hazard\cr
}
}
\description{
Given time to event data, a set of cuts and a sequence of penalty values, returns the adaptive ridge or the ridge estimator of the log-hazard.
}
\details{
The adaptive-ridge algorithm is run for the sequence of penalty values \code{pen}. The final adaptive-ridge estimator \code{final.a} is found
by minimising a Bayesian Information Criteria (BIC). The \code{res} list returned from the function contains all the
estimators (adaptive-ridge or ridge) for each penalty value. Write \code{res[[pos]]} to get the estimator for the
penalty corresponding to \code{pen[pos]}. Use the option AR to choose between the ridge or adaptive-ridge estimator.
For the ridge estimator only res is not null.Confidence intervals are computed for the final adaptive-ridge hazard estimator.
}
\examples{
set.seed(45)
n=400
cuts=c(20,40,50,70)
alpha=c(0,0.05,0.1,0.2,0.4)/10
time=rsurv(n,cuts,alpha) #generate true data from the pch model
censoring=runif(n,min=70,max=90)
time=pmin(time,censoring) #observed times
delta=time<censoring #gives 62\% of observed data on average

#The adaptive ridge estimator
fit=arpchsurv(time,delta,verbose=TRUE,cuts=1:100,CI=TRUE)

fit$final.cuts
exp(fit$final.a)

plot(fit)
plot(fit,pos=fit$pos,CI=TRUE)
points(c(0,c(20,40,50,70),max(time)),c(0,alpha),type="S",lwd=2,xlim=c(10,80),lty=2,col="red") #the true hazard function

fitsurv=pchsurv(time,delta,cuts=fit$final.cuts)
plot(fitsurv,CI=TRUE)
seqtime=seq(0,100,by=0.1)
lines(seqtime,exp(-pchcumhaz(seqtime,cuts,alpha)),type="l",col="red") #the true survival function

#The ridge estimator
fitR=arpchsurv(time,delta,verbose=TRUE,cuts=1:100,AR=FALSE)
plot(fitR)
plot(fitR,pos=50,main="")
lines(fitR,pos=80,col="blue")
points(c(0,c(20,40,50,70),max(time)),c(0,alpha),type="S",lwd=2,xlim=c(10,80),lty=2,col="red") #the true hazard function
}
\seealso{
Other pchsurv functions: \code{\link{bootpchsurv}},
  \code{\link{mlepchsurv}}, \code{\link{pchsurv}},
  \code{\link{rsurv}}
}
\concept{pchsurv functions}
\keyword{pchsurv}
