% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpchsurv.R
\name{bootpchsurv}
\alias{bootpchsurv}
\title{Bootstrap estimator of the adaptive piecewise constant hazard model}
\usage{
bootpchsurv(time, status, cuts, seqtime = NULL, M = 100,
  verbose = TRUE, tol = 1e-07, itermax = 1e+05,
  pen = exp(seq(log(0.1), log(1000), length = 100)), w = rep(1,
  length(cuts)), a = rep(0, length(cuts) + 1))
}
\arguments{
\item{time}{the observed time data.}

\item{status}{status of the time data. TRUE for true time and FALSE for censored time.}

\item{cuts}{a sequence of cuts. Default to NULL which corresponds to the exponential model.}

\item{seqtime}{a time sequence where the survival function is estimated.}

\item{M}{the number of bootrapped samples.}

\item{tol}{the tolerance parameter for convergence of the algorithm. Default to 1e-7.}

\item{itermax}{the maximum number of iterations. If the algorithm has not converged before itermax iterations then the algorithm exits the program. Default to 1e+5.}

\item{pen}{a sequence of penalty values.}

\item{w}{the w sequence values for the adaptive ridge algorithm at the initialization step. Default to 1. Should be of the size of the cuts.}

\item{a}{the log-hazard value at the initialization step. Default to the unpenalized log-hazard estimator.}
}
\value{
\tabular{lll}{
\code{Result} \tab  \code{ } \tab a matrix containing the cumulative hazard estimates for each bootstrap sample\cr
\tab \tab \cr
\code{medSurv} \tab \code{ }  \tab the median estimator obtained from all bootstrap samples\cr
\tab \tab \cr
\code{CIleft} \tab  \code{ } \tab the left confidence intervals obtained from the bootstrap samples\cr
\tab \tab \cr
\code{CIright} \tab \code{ }  \tab the right confidence intervals obtained from the bootstrap samples\cr
}
}
\description{
Given time to event data and a set of cuts return the survival estimator from maximum likelihood estimation in the piecewise constant hazard model
}
\details{
The bootstrap procedure is performed by sampling on the initial sample. A total of \code{M} bootstrap samples are generated.
A new estimator is constructed by taking the median of all survival estimate for all bootrapp samples. Confidence intervals are computed
using the bootstrap samples.
}
\examples{
n=400
cuts=c(20,40,50,70)
alpha=c(0,0.05,0.1,0.2,0.4)/10
time=rsurv(n,cuts,alpha) #generate true data from the pch model
censoring=runif(n,min=70,max=90)
time=pmin(time,censoring) #observed times
delta=time<censoring #gives 62\% of observed data on average
Result=bootpchsurv(time,delta,cuts=1:100,M=20)
plot(Result)

#The adaptive ridge estimator
fit=arpchsurv(time,delta,verbose=TRUE,cuts=1:100,CI=TRUE)
fitsurv=pchsurv(time,delta,cuts=fit$final.cuts)
lines(fitsurv,CI=TRUE,col="blue")
seqtime=seq(0,100,by=0.1)
lines(seqtime,exp(-pchcumhaz(seqtime,cuts,alpha)),type="l",col="red") #the true survival function


}
\seealso{
Other pchsurv functions: \code{\link{arpchsurv}},
  \code{\link{mlepchsurv}}, \code{\link{pchsurv}},
  \code{\link{rsurv}}
}
\concept{pchsurv functions}
\keyword{pchsurv}
