% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.pchsurv.R
\name{loglik.pchsurv}
\alias{loglik.pchsurv}
\title{Likelihood function of a piecewise constant hazard model}
\usage{
loglik.pchsurv(time, status, cuts, a, weights = NULL)
}
\arguments{
\item{time}{the observed time data.}

\item{status}{status of the time data. TRUE for true time and FALSE for censored time.}

\item{cuts}{a sequence of cuts. Default to NULL which corresponds to the exponential model.}

\item{a}{sequence of log-hazard values. Should be of length equal to length(cuts)+1.}

\item{weights}{an optional weight sequence. Default to NULL.}
}
\description{
Given time to event data, a set of cuts and a log-hazard value, compute the maximum likelihood function
}
\examples{
n=400
cuts=c(20,40,50,70)
alpha=c(0,0.05,0.1,0.2,0.4)/10
time=rsurv(n,cuts,alpha) #generate true data from the pch model
censoring=runif(n,min=70,max=90)
time=pmin(time,censoring) #observed times
delta=time<censoring #gives 62\% of observed data on average
result=mlepchsurv(time,delta,cuts)
loglik.pchsurv(time,delta,cuts,result$a)
}
\keyword{pchsurv}
