% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlepchsurv.R
\name{mlepchsurv}
\alias{mlepchsurv}
\title{Maximum likelihood estimation from the piecewise constant hazard model}
\usage{
mlepchsurv(time, status, cuts = NULL, weights = NULL, CI = TRUE,
  alphaCI = 0.05, logtransf = TRUE)
}
\arguments{
\item{time}{observed time data.}

\item{status}{status of the time data. TRUE for true time and FALSE for censored time.}

\item{cuts}{the sequence of cuts. Default to NULL which corresponds to the exponential model.}

\item{weights}{an optional weight sequence. Default to NULL.}

\item{CI}{should the confidence intervals be computed? Default to TRUE.}

\item{alphaCI}{the value of alpha for 1-alpha confidence intervals. Default to 0.05.}

\item{logtransf}{should the confidence intervals be computed using the log-transformtation? Default to TRUE.}
}
\value{
\tabular{lll}{
\code{a} \tab  \code{ } \tab the estimated log-hazard\cr
\tab \tab \cr
\code{hazard} \tab \code{ }  \tab the estimated hazard\cr
\tab \tab \cr
\code{A} \tab  \code{ } \tab the number of observed events between each cut\cr
\tab \tab \cr
\code{B} \tab \code{ }  \tab the total time at risk between each cut\cr
\tab \tab \cr
\code{CIleft} \tab  \code{ } \tab the left confidence intervals\cr
\tab \tab \cr
\code{CIright} \tab \code{ }  \tab the right confidence intervals\cr
}
}
\description{
Given time to event data and a set of cuts return the maximum likelihood estimator of the log-hazard from the piecewise constant hazard model.
}
\details{
The maximum likelihood estimator is computed from the two exhaustive statistics O_k=sum_i Delta_i 1(c_{(k-1)}<T_i<=c_k) and R_k=sum_i (min(c_k,T_i)-c_{(k-1)}) 1(T_i>=c_{(k-1)}).
It is equal to O_k/R_k on each cut (c_{(k-1)},c_{k}]. They are called A and B in what follows.
}
\examples{
n=400
cuts=c(20,40,50,70)
alpha=c(0,0.05,0.1,0.2,0.4)/10
time=rsurv(n,cuts,alpha) #generate true data from the pch model
censoring=runif(n,min=70,max=90)
time=pmin(time,censoring) #observed times
delta=time<censoring #gives 62\% of observed data on average
mlepchsurv(time,delta,cuts)
}
\seealso{
Other pchsurv functions: \code{\link{arpchsurv}},
  \code{\link{bootpchsurv}}, \code{\link{pchsurv}},
  \code{\link{rsurv}}
}
\concept{pchsurv functions}
\keyword{pchsurv}
