% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchsurv.R
\name{pchsurv}
\alias{pchsurv}
\title{Estimated survival function from the piecewise constant hazard model}
\usage{
pchsurv(time, status, cuts, seqtime = NULL, CI = TRUE,
  alphaCI = 0.05)
}
\arguments{
\item{time}{the observed time data.}

\item{status}{status of the time data. TRUE for true time and FALSE for censored time.}

\item{cuts}{a sequence of cuts. Default to NULL which corresponds to the exponential model.}

\item{seqtime}{a time sequence.}

\item{CI}{should the confidence intervals be computed? Default to TRUE.}

\item{alphaCI}{the value of alpha for 1-alpha confidence intervals. Default to 0.05.}
}
\value{
\tabular{lll}{
\code{surv} \tab  \code{ } \tab the estimated survival function\cr
\tab \tab \cr
\code{CIleft} \tab  \code{ } \tab the left confidence intervals of the survival function\cr
\tab \tab \cr
\code{CIright} \tab \code{ }  \tab the right confidence intervals of the survival function\cr
}
}
\description{
Given time to event data and a set of cuts return the survival estimator from maximum likelihood estimation in the piecewise constant hazard model
}
\details{
The survival estimator is computed from the maximum likelihood estimator in the piecewise constant hazard model. It is computed for each value of \code{seqtime}.
}
\examples{
n=400
cuts=c(20,40,50,70)
alpha=c(0,0.05,0.1,0.2,0.4)/10
time=rsurv(n,cuts,alpha) #generate true data from the pch model
censoring=runif(n,min=70,max=90)
time=pmin(time,censoring) #observed times
delta=time<censoring #gives 62\% of observed data on average
seqtime=seq(0,100,by=10)
result=pchsurv(time,delta,cuts,seqtime)
result
plot(result,CI=TRUE)
}
\seealso{
Other pchsurv functions: \code{\link{arpchsurv}},
  \code{\link{bootpchsurv}}, \code{\link{mlepchsurv}},
  \code{\link{rsurv}}
}
\concept{pchsurv functions}
\keyword{pchsurv}
